Updates project onboarding (#1)
## Why

This PR updates the projects-onboarding doc so that it contains more
information about how AI agents can access accurate information about
Expo, the SDK, and services.

## Test plan

Use these contents inside of an AGENTS.md file, then chat with an LLM
and make sure that the agent respects these instructions and that the
output is reasonable.

If there's a better way to evaluate the effectiveness of these changes,
id love to learn more!